Move FastAPI route/router classes out of app.py
Keep App focused on lifecycle/setup. PulseAPIRoute, PulseAPIRouter, and
PulseFastAPI now live in api_router.py. Lazy-import PULSE_CONTEXT to
avoid the app ↔ context cycle.
