+++ packages/pulse/python/README.md
@@ -31,6 +31,7 @@ Server-driven UI model: Python components render to VDOM, synced to React via We
 ```
 src/pulse/
 ├── app.py              # Main App class, FastAPI + Socket.IO setup
+├── api_router.py       # FastAPI route/router classes (user API + /_pulse)
 ├── channel.py          # Bidirectional real-time channels
 ├── routing.py          # Route/Layout definitions, URL matching
 ├── vdom.py             # VDOM node types (Element, Component, Node)
